Extract hooks to a listener.
