+++ src/Resources/contao/config/config.php
@@ -1,42 +1,21 @@
 <?php
 
 /**
- * htaccess Generator
- * Copyright (C) 2011 Tristan Lins
+ * MetaPalettes for the Contao Open Source CMS
  *
- * Extension for:
- * Contao Open Source CMS
- * Copyright (C) 2005-2011 Leo Feyer
- *
- * Formerly known as TYPOlight Open Source CMS.
- *
- * This program is free software: you can redistribute it and/or
- * modify it under the terms of the GNU Lesser General Public
- * License as published by the Free Software Foundation, either
- * version 3 of the License, or (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
- * Lesser General Public License for more details.
- *
- * You should have received a copy of the GNU Lesser General Public
- * License along with this program. If not, please visit the Free
- * Software Foundation website at <http://www.gnu.org/licenses/>.
- *
- * PHP version 5
- * @copyright  InfinitySoft 2011
- * @author     Tristan Lins <tristan.lins@infinitysoft.de>
- * @package    MetaPalettes
- * @license    LGPL
- * @filesource
+ * @link      https://github.com/bit3/contao-meta-palettes
+ * @copyright 2013-2014 bit3 UG
+ * @copyright 2015-2017 Contao Community Alliance
+ * @author    Tristan Lins <tristan.lins@infinitysoft.de>
+ * @author    David Molineus <david.molineus@netzmacht.de>
+ * @package   MetaPalettes
+ * @license   LGPL-3.0+
  */
-
 
 /**
  * Hooks
  */
-$GLOBALS['TL_HOOKS']['loadDataContainer'][] = array('ContaoCommunityAlliance\MetaPalettes\MetaPalettes', 'generatePalettes');
+$GLOBALS['TL_HOOKS']['loadDataContainer'][] = ['cca.meta_palettes.listener.hook_listener', 'generatePalettes'];
 
 $GLOBALS['TL_EVENTS']['dc-general.factory.build-data-definition'][] = array(
     'ContaoCommunityAlliance\MetaPalettes\MetaPalettesBuilder::process',
